backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
supabase==2.10.0
groq==0.12.0
sentence-transformers==3.3.1
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
